Add GitCondDB as an external
To assist in testing an Art Service for GitCondDB, import it as an
external project as we don't have a UPS product for it. Include
deps libgit2, libssh2, fmt and nlohmann_json as externals, with
everything installing into a subdirectory of the build area.

Declare mock imported targets for GitCondDB and nlohmann_json so that
no changes to build of code will be needed once UPS/Spack packages
are available.