Import resource files from upstream Falaise
To test resource setup, we need the files to work with. These exist in
the upstream Falaise repository, so for simplicity copy the files in
from that repo into the same location. It's purely a dump to check
functionality. For now we keep them next to service that uses them.